JSON AST: extract PublicAST.Pattern
