+++ elm-format-lib/elm-format-lib.cabal
@@ -28,6 +28,7 @@ common common-options
         LambdaCase
         OverloadedStrings
         ScopedTypeVariables
+        FlexibleInstances
 
     hs-source-dirs:
         src
@@ -88,6 +89,8 @@ library
         CommandLine.Program
         CommandLine.ResolveFiles
         CommandLine.TransformFiles
+        ElmFormat.AST.PublicAST
+        ElmFormat.AST.Shared
         ElmFormat.ImportInfo
         ElmFormat.KnownContents
         ElmFormat.Parse
@@ -115,6 +118,8 @@ test-suite efl-tests
         AST.Structure
         AST.V0_16
         CommandLine.InfoFormatter
+        ElmFormat.AST.PublicAST
+        ElmFormat.AST.Shared
         ElmFormat.ImportInfo
         ElmFormat.KnownContents
         ElmFormat.Parse
