+++ elm-format-lib/src/AST/V0_16.hs
@@ -6,20 +6,16 @@
 {-# LANGUAGE GADTs #-}
 {-# LANGUAGE PolyKinds #-}
 
-module AST.V0_16 where
+module AST.V0_16 (module AST.V0_16, module ElmFormat.AST.Shared) where
 
 import Data.Bifunctor
 import Data.Coapplicative
 import Data.Foldable
 import Data.Functor.Const
 import Data.Functor.Compose
-import Data.Int (Int64)
 import qualified Data.Indexed as I
-import qualified Data.Maybe as Maybe
 import qualified Cheapskate.Types as Markdown
-
-
-type List a = [a]
+import ElmFormat.AST.Shared
 
 
 newtype ForceMultiline =
@@ -28,52 +24,6 @@ newtype ForceMultiline =
 
 instance Semigroup ForceMultiline where
     (ForceMultiline a) <> (ForceMultiline b) = ForceMultiline (a || b)
-
-
-data LowercaseIdentifier =
-    LowercaseIdentifier String
-    deriving (Eq, Ord, Show)
-
-
-data UppercaseIdentifier =
-    UppercaseIdentifier String
-    deriving (Eq, Ord, Show)
-
-
-data SymbolIdentifier =
-    SymbolIdentifier String
-    deriving (Eq, Ord, Show)
-
-
-data Comment
-    = BlockComment (List String)
-    | LineComment String
-    | CommentTrickOpener
-    | CommentTrickCloser
-    | CommentTrickBlock String
-    deriving (Eq, Ord, Show)
-type Comments = List Comment
-
-eolToComment :: Maybe String -> Comments
-eolToComment eol =
-    Maybe.maybeToList (fmap LineComment eol)
-
-
-data Commented c a =
-    C c a
-    deriving (Eq, Ord, Functor, Show) -- TODO: is Ord needed?
-
-instance Coapplicative (Commented c) where
-    extract (C _ a) = a
-    {-# INLINE extract #-}
-
-type C1 l1 a = Commented Comments a
-type C2 l1 l2 a = Commented (Comments, Comments) a
-type C3 l1 l2 l3 a = Commented (Comments, Comments, Comments) a
-
-type C0Eol a = Commented (Maybe String) a
-type C1Eol l1 a = Commented (Comments, Maybe String) a
-type C2Eol l1 l2 a = Commented (Comments, Comments, Maybe String) a
 
 
 {-| This represents a list of things separated by comments.
@@ -217,49 +167,10 @@ instance Foldable (NameWithArgs name) where
     foldMap f (NameWithArgs _ args) = foldMap (f . extract) args
 
 
-data IntRepresentation
-  = DecimalInt
-  | HexadecimalInt
-  deriving (Eq, Show)
-
-
-data FloatRepresentation
-  = DecimalFloat
-  | ExponentFloat
-  deriving (Eq, Show)
-
-
-data StringRepresentation
-    = SingleQuotedString
-    | TripleQuotedString
-    deriving (Eq, Show)
-
-
-data LiteralValue
-    = IntNum Int64 IntRepresentation
-    | FloatNum Double FloatRepresentation
-    | Chr Char
-    | Str String StringRepresentation
-    | Boolean Bool
-    deriving (Eq, Show)
-
-
-data Ref ns
-    = VarRef ns LowercaseIdentifier
-    | TagRef ns UppercaseIdentifier
-    | OpRef SymbolIdentifier
-    deriving (Eq, Ord, Show, Functor)
-
-
 data TypeConstructor ctorRef
     = NamedConstructor ctorRef
     | TupleConstructor Int -- will be 2 or greater, indicating the number of elements in the tuple
     deriving (Eq, Show, Functor)
-
-
-data UnaryOperator =
-    Negative
-    deriving (Eq, Show)
 
 
 data BinopsClause varRef expr =
